+++ site/webapp/controller/helloVue.js
@@ -3,8 +3,8 @@
  * @tagline         Demo Vue.js SPA Controller
  * @description     Example of Vue.js Single Page Application integration
  * @file            site/webapp/controller/helloVue.js
- * @version         1.6.31
- * @release         2026-03-20
+ * @version         1.6.32
+ * @release         2026-03-21
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/