+++ webapp/tests/unit/utils/context-extensions.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for ContextExtensions (W-014)
  * @description     Tests Handlebars context extension system
  * @file            webapp/tests/unit/utils/context-extensions.test.js
- * @version         1.6.31
- * @release         2026-03-20
+ * @version         1.6.32
+ * @release         2026-03-21
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/